+++ CHANGELOG.md
@@ -15,13 +15,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Batch dimension userguide
 - Parallel inference example
 - Perturbation method section in userguide
-- WeatherBench Climatology data source
+- WeatherBench Climatology and ERA5 data source
 - Added `datasource_to_file` utility function
 - Add lagged ensemble perturbation method
 - Add ACC and CRPS metrics
 - Added the ability to reload zarr and netcdf backends
 - Added the ability to read from an IOBackend
 - Add spread/skill ratio
+- Added rank histogram
+- Added reduction_dimensions as required property in statistics
+  and metrics API.
 
 ### Changed